+++ CHANGELOG.md
@@ -8,12 +8,12 @@ how this file is formatted and how the process works.
 ### Added
 - Vietnam country code for SAW
 - THH: check in/out time info
+- Poplidays: skip properties with zero `max_guest`
 
 ## [1.3.1] - 2016-11-18
 ### Added
 - Laos country code for SAW
 - WayToStay city tourist tax and late checkin fee to description append
-- Poplidays: skip properties with zero `max_guest`
 
 ### Fixed
 - THH: min_stay = max(min_nights)